Modify template to work with partial profile
This is to work around a CI issue whereby each profile is checked in turn. In the case of a partial profile (one which needs to be combined with a another profile) the CI test was failing as the template did not handle some undefined sections. This should have no impact when the exporters are correctly used from the command line.